Fixes warnings

+++ src/branch.cpp
@@ -172,7 +172,7 @@ void Branch::print(std::string level)
 std::vector<RowID> Branch::selectAll()
 {
 	std::vector<RowID> rowIds;
-	for (int i = 0; i < block.size(); i++) {
+	for (size_t i = 0; i < block.size(); i++) {
 		std::vector<RowID> leftIds = block[i].getLeft()->selectAll();
 		rowIds.insert(rowIds.end(), leftIds.begin(), leftIds.end());
 		if(i == block.size() - 1)
